refactor: move eslint package to @n8n
