+++ packages/@n8n/eslint-config/base.js
