+++ packages/@n8n/eslint-config/frontend.js
