+++ packages/@n8n/eslint-config/jest.config.js
