+++ packages/@n8n/eslint-config/shared.js
