+++ packages/@n8n/vitest-config/frontend.mjs
