+++ packages/design-system/.eslintrc.js
@@ -1,9 +1,13 @@
-const { createFrontendEslintConfig } = require('@n8n/frontend-eslint-config');
+const sharedOptions = require('@n8n/eslint-config/shared');
 
 /**
  * @type {import('@types/eslint').ESLint.ConfigData}
  */
-module.exports = createFrontendEslintConfig(__dirname, {
+module.exports = {
+	extends: ['@n8n/eslint-config/frontend'],
+
+	...sharedOptions(__dirname, 'frontend'),
+
 	rules: {
 		// TODO: Remove these
 		'import/no-default-export': 'warn',
@@ -40,4 +44,4 @@ module.exports = createFrontendEslintConfig(__dirname, {
 			},
 		},
 	],
-});
+};
