+++ packages/frontend/@n8n/chat/.eslintrc.cjs
@@ -1,10 +1,10 @@
-const sharedOptions = require('@n8n_io/eslint-config/shared');
+const sharedOptions = require('@n8n/eslint-config/shared');
 
 /**
  * @type {import('@types/eslint').ESLint.ConfigData}
  */
 module.exports = {
-	extends: ['@n8n_io/eslint-config/frontend'],
+	extends: ['@n8n/eslint-config/frontend'],
 
 	...sharedOptions(__dirname, 'frontend'),
 };
