+++ packages/frontend/@n8n/composables/vite.config.ts
@@ -1,4 +1,4 @@
 import { defineConfig, mergeConfig } from 'vite';
-import { vitestConfig } from '@n8n/frontend-vitest-config';
+import { vitestConfig } from '@n8n/vitest-config/frontend';
 
 export default mergeConfig(defineConfig({}), vitestConfig);
