--- packages/frontend/tooling/eslint-config/index.cjs
@@ -1,14 +1,0 @@
-const sharedOptions = require('@n8n_io/eslint-config/shared');
-
-/**
- * @param {string} path
- * @param {import('@types/eslint').ESLint.ConfigData} config
- * @returns {import('@types/eslint').ESLint.ConfigData}
- */
-module.exports.createFrontendEslintConfig = function (path, config = {}) {
-	return {
-		extends: ['@n8n_io/eslint-config/frontend'],
-		...sharedOptions(path, 'frontend'),
-		...config,
-	};
-};
