--- packages/frontend/tooling/eslint-config/index.d.ts
@@ -1,3 +1,0 @@
-import type { ESLint } from '@types/eslint';
-
-export function createFrontendEslintConfig(path: string): ESLint.ConfigData;
