--- packages/frontend/tooling/eslint-config/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "@n8n/frontend-eslint-config",
-  "version": "1.1.0",
-  "type": "module",
-  "dependencies": {
-    "@n8n_io/eslint-config": "workspace:*",
-    "eslint-plugin-n8n-local-rules": "^1.0.0"
-  },
-  "files": [
-    "index.cjs"
-  ],
-  "main": "./index.cjs",
-  "module": "./index.cjs",
-  "exports": {
-    ".": {
-      "import": "./index.cjs",
-      "require": "./index.cjs",
-      "types": "./index.d.ts"
-    },
-    "./*": "./*"
-  },
-  "license": "See LICENSE.md file in the root of the repository"
-}
