+++ pnpm-workspace.yaml
@@ -1,7 +1,6 @@
 packages:
   - packages/*
   - packages/@n8n/*
-  - packages/@n8n_io/*
   - packages/frontend/**
   - cypress
